+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Views
 Extensions
 
 - `View.getNestedSize(_:)` method is added, that returns `CGSize` of the nested `View`
+- `UIHostingController.overrideBehaviors(_:)` method is added, that overrides select behaviors
 
 ### [5.0.0(85)](https://github.com/VakhoKontridze/VCore/releases/tag/5.0.0) — *2023 10 08*